Move metrics types fromto src/lib/{charts,metrics}
