+++ src/lib/metric-data-getters/cloudfront.ts
@@ -8,7 +8,7 @@ import {
   IMetricDataGetter,
   IMetricSerie,
   MetricName,
-} from '../charts';
+} from '../metrics';
 
 import { CloudWatchMetricDataGetter, ICloudWatchMetric } from './cloudwatch';
