+++ src/lib/metric-data-getters/elasticache.ts
@@ -10,7 +10,7 @@ import {
   IMetricDataGetter,
   IMetricSerie,
   MetricName,
-} from '../charts';
+} from '../metrics';
 
 import { CloudWatchMetricDataGetter, ICloudWatchMetric } from './cloudwatch';
